feat: add shared base agent
